+++ crates/miden-objects/src/transaction/tx_args.rs
@@ -2,9 +2,10 @@ use alloc::collections::{BTreeMap, BTreeSet};
 use alloc::sync::Arc;
 use alloc::vec::Vec;
 
+use miden_crypto::dsa::rpo_falcon512::PublicKey;
 use miden_crypto::merkle::InnerNodeInfo;
 
-use super::{AccountInputs, Felt, Word};
+use super::{AccountInputs, Felt, Hasher, Word};
 use crate::note::{NoteId, NoteRecipient};
 use crate::utils::serde::{
     ByteReader,
@@ -185,6 +186,17 @@ impl TransactionArgs {
         ];
 
         self.advice_inputs.extend(AdviceInputs::default().with_map(new_elements));
+    }
+
+    /// Adds the `signature` corresponding to `public_key` on `message` to the advice inputs' map.
+    ///
+    /// The advice inputs' map is extended with the following key:
+    ///
+    /// - hash(public_key, message) |-> signature.
+    pub fn add_signature(&mut self, public_key: PublicKey, message: Word, signature: Vec<Felt>) {
+        self.advice_inputs
+            .map
+            .insert(Hasher::merge(&[public_key.into(), message]), signature);
     }
 
     /// Populates the advice inputs with the specified note recipient details.